# Use package details from extracted files