… RELEASE_NOTES.md
… src/Paket.Core/Dependencies/NuGetLocal.fs
… src/Paket.Core/Dependencies/NuGetV3.fs
… src/Paket.Core/Dependencies/PackageResolver.fs
… src/Paket/AssemblyInfo.fs
… tests/Paket.Tests/NuGetLocal/NuGetLocalSpecs.fs