two improvements:

1) instead of getting package details (meta data) from NuGet API we first try to look if we already have the package in some local cache. If so then we extract it's nuspec and use the meta data from within. This saves a lot of request in early stages after SDK installation. The main packages of .NET Standard and asp.net are often already in the NuGet Fallback folder.

2) in our lookahead routine we didn't deduplicate requests. We now do that and also we limit to 3 versions lookahead instead of 10.