@matthid 2019-11-04
Was there any specific need to do this?
Just two notes (all in all looks good but it is a bit hard to see the changes due to the moved code):
- Should we write the request cache file when we already use the nuspec file? I mean this is basically already a cache so no need to still write a second one.
- Have you some numbers regarding the perf? I'm asking because I wouldn't wonder if on slow hard drives doing the request could be faster, maybe? But I agree that overall this should improve things.