@forki 2019-11-04
In @isaac_abraham's example it went from 18s to 8s for me. The double caching is fast because we do not have to look in multiple folders.