@isaacabraham 2019-11-05
@mattid for me, just doing an install on Saturn **with every package already on my local hard disk** took 58 seconds with Paket vs 5 seconds with NuGet.