@matthid 2019-11-05
@isaacabraham What about with and without this PR?
In any case the changes look good to me (as I said already), just was interested about the background