@isaacabraham 2019-11-05
@matthid yep, i'm waiting for nuget to finish indexing and will then test :-)