@isaacabraham 2019-11-05
## Before:
Performance:
 - Resolver: 51 seconds (1 runs)
    - Runtime: 804 milliseconds
    - Blocked (retrieving package details): 45 seconds (63 times)
    - Blocked (retrieving package versions): 5 seconds (8 times)
    - Not Blocked (retrieving package details): 72 times
    - Not Blocked (retrieving package versions): 127 times
 - Average Request Time: 109 milliseconds
 - Number of Requests: 459
 - Runtime: 53 seconds

## After
Performance:
 - Resolver: 20 seconds (1 runs)
    - Runtime: 593 milliseconds
    - Blocked (retrieving package details): 13 seconds (135 times)
    - Blocked (retrieving package versions): 7 seconds (10 times)
    - Not Blocked (retrieving package versions): 125 times
 - Disk IO: 1 second
 - Average Request Time: 147 milliseconds
 - Number of Requests: 135
 - Runtime: 22 seconds