@isaacabraham 2019-11-05
## After + v3 API
 - Resolver: 7 seconds (1 runs)
    - Runtime: 546 milliseconds
    - Blocked (retrieving package details): 4 seconds (135 times)
    - Blocked (retrieving package versions): 2 seconds (9 times)
    - Not Blocked (retrieving package versions): 126 times
 - Disk IO: 187 milliseconds
 - Average Request Time: 49 milliseconds
 - Number of Requests: 136
 - Runtime: 8 seconds