@isaacabraham 2019-11-05
Awesome. I do wonder if Nuget does anything else differently?

Also shouldn't default paket init use v3 api now? It's way faster.