@isaacabraham 2019-11-05
What actually is the blocked count anyway