@matthid 2019-11-05
@isaacabraham We have some logic to "preload" some requests. Not blocked means this logic is reasonable good because we already have this "prefetched" and the logic can continue. "Blocked" means the resolver stops doing anything and waiting for a request to finish, this always happens because we can never have a perfect preloading logic, but often it is quite predictable (for example we need for all packages in paket.dependencies all available version obviously).