@matthid 2019-11-05
Maybe this PR should have changed the count because now everything seems to count as "blocked" even though we asked the local cache. @forki