@forki 2019-11-05
yeah I guess there is just no prefetching happening anymore because shit is faster